lkl: support mach-o binary for alternate host environment
This commit introduces an alternate posix environment on macOS.
Since the object format used on this OS, mach-o is different from
others, which uses ELF, we need to take care of features which
assume ELF.

So far, section format and lack of linker script are under consideration
of this support.  The more we test, the more we find other differences we
guess.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>